Refactor chat runtime architecture to integrate new runtime registry
- Replaced `MountedRuntimes` with `RuntimeSlots` in `ChatProviders` to align with the new runtime registry structure.
- Updated `ChatRouteHost` to utilize the new `useRuntime` and `useRuntimeActions` hooks from the `runtime-registry`, enhancing runtime management.
- Removed the deprecated `chat-runtime` module and its associated documentation, streamlining the codebase and improving clarity on runtime handling.
- Introduced new `RuntimeRegistryProvider` and related components to manage runtime lifecycles more effectively, ensuring better separation of concerns in the chat architecture.
- Revised documentation to reflect the changes in runtime management and clarify the new architecture for future development.
